@@ -6,36 +6,32 @@ using CommandLine;
 namespace ExcelToJson
 {
     /// <summary>
-    /// 命令行参数定义
+    ///     命令行参数定义
     /// </summary>
     public sealed class Options
     {
-        public static Options Default = new Options();
+        public static Options Default = new();
 
         public Encoding Encoding => new UTF8Encoding(false);
 
-        [Option('e', "excel_path")] public string ExcelPath { get; set; }
-        [Option('j', "json_path")] public string JsonPath { get; set; }
-        [Option('c', "script_path")] public string ScriptPath { get; set; }
-        [Option('t', "script_template_path")] public string ScriptTemplate { get; set; }
+        [Option("excel_path")]
+        public string ExcelPath { get; set; }
 
-        [Option('l', "script_load_json")] 
-        public string ScriptLoadJsonPath { get; set; } = "#json_name";
+        [Option("json_path")]
+        public string JsonPath { get; set; }
 
-        [Option('b',"bson")]
-        public int IsBinaryJson { get; set; }
+        [Option("script_path")]
+        public string ScriptPath { get; set; }
 
-        public string JsonFileExtension => IsBinaryJson == 1 ? ".bytes" : ".json";
+        [Option("script_template_path")]
+        public string ScriptTemplate { get; set; }
 
         public void Check()
         {
-            ExcelPath = Path.GetFullPath(ExcelPath);
-            ScriptTemplate = Path.GetFullPath(ScriptTemplate);
-            if (Path.HasExtension(ScriptLoadJsonPath))
-                ScriptLoadJsonPath = ScriptLoadJsonPath.Replace(Path.GetExtension(ScriptLoadJsonPath), "");
-
             if (!Directory.Exists(ExcelPath))
+            {
                 Console.WriteLine($"{ExcelPath} not exist");
+            }
 
             if (string.IsNullOrEmpty(ScriptPath))
             {
@@ -43,16 +39,15 @@ namespace ExcelToJson
             }
             else
             {
-                ScriptPath = Path.GetFullPath(ScriptPath);
+                if (!File.Exists(ScriptTemplate))
+                {
+                    Console.WriteLine($"ScriptTemplatePath:{ScriptTemplate} not exist");
+                }
             }
 
             if (string.IsNullOrEmpty(JsonPath))
             {
                 Console.WriteLine("JsonPath is null");
-            }
-            else
-            {
-                JsonPath = Path.GetFullPath(JsonPath);
             }
         }
     }
